\name{babyNames}
\alias{babyNames}
\docType{data}
\title{
Data on baby names
}
\description{
Data on baby names from ssa.gov
}
\usage{data(babyNames)}
\format{
  A data frame with 91196 observations on the following 3 variables.
  \describe{
    \item{\code{name}}{a character vector}
    \item{\code{gender}}{a character vector}
    \item{\code{total}}{a numeric vector}
  }
}
\source{
http://www.ssa.gov/OACT/babynames/names.zip
}
\examples{
data(babyNames)
}
\keyword{datasets}
