% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-genderdata-package.R
\name{check_genderdata_package}
\alias{check_genderdata_package}
\title{Check whether to install data for gender function and install if necessary}
\usage{
check_genderdata_package()
}
\description{
If the genderdata package is not installed,
or it is out of date (earlier than version 0.5.0),
install it from ropensci after checking with the user.
}
