\name{gender-package}
\alias{gender-package}
\docType{package}
\title{
Guess gender of individual based on first name
}
\description{
Guess gender of individual based on first name
}
\details{
\tabular{ll}{
Package: \tab gender\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03-29\cr
License: \tab GPL\cr
}
gender() is the main function.
}
\author{
Scott Kostyshak

Maintainer: Scott Kostyshak <skostysh@princeton.edu>
}
\keyword{ package }
