% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gender_demo}
\alias{gender_demo}
\title{Demo finding the gender of first names using SSA data}
\usage{
gender_demo(name, years, certainty)
}
\arguments{
\item{name}{A character string of a first name. Case insensitive.}

\item{years}{This argument can be either a single year or a range of years in
the form \code{c(1880, 1900)}. If no value is specified, then the names
will be looked up for the period 1932 to 2012. If a year or range of years
is specified, then the names will be looked up for that period. Dates may
range from 1880 to 2012. For years before 1930, the IPUMS method is
probably better.}

\item{certainty}{A boolean value, which determines whether or not to return
the proportion of male and female uses of names in addition to determining
the gender of names.}
}
\description{
This internal function implements the \code{method = "demo"} option of
\code{\link{gender}}. See that function for documentation.
}

