% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gender-genderize.R
\name{gender_genderize}
\alias{gender_genderize}
\title{Find the gender of first names using Genderize.io API}
\usage{
gender_genderize(name, certainty)
}
\arguments{
\item{name}{A character string of a first name.}

\item{certainty}{A boolean value, which determines whether or not to return
the proportion of male and female uses of names in addition to determining
the gender of names.}
}
\value{
A list or (for multiple names) a list of lists containing the name
  property and the predicted gender property, along with the proportion of
  the uses of the name that is male and female.
}
\description{
This internal function implements the \code{method = "genderize"} option of
\code{\link{gender}}. See that function for documentation.
}

