% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gender_kantrowitz}
\alias{gender_kantrowitz}
\title{Find the gender of frst names using Kantrowitz names corpus}
\usage{
gender_kantrowitz(data)
}
\arguments{
\item{data}{A character string of a first name or a data frame with a column
named \code{name} with a character vector containing first names. The names
must all be lowercase.}
}
\description{
This internal function implements the \code{method = "kantrowitz"} option of
\code{\link{gender}}. See that function for documentation.
}

