% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_correction_factors}
\alias{get_correction_factors}
\title{Calculate the correction factors for a year or range of years}
\usage{
get_correction_factors(years)
}
\arguments{
\item{years}{The range of years. This value will be passed to it by the
gender_ssa function.}

\item{names_df}{The data frame of names and genders used inside the function.
Defaults to the U.S. Social Security Administration's database of baby
names provided by the \code{babynames} package.}
}
\description{
The SSA data is skewed by gender, especially for years before 1935. This
internal function figures out the factor by which the gender ratio should be
multiplied in order to assume that the ratio of male to female births for
those years was 1:1.
}

