% Generated by roxygen2 (4.0.2): do not edit by hand
\name{install_genderdata_package}
\alias{install_genderdata_package}
\title{Install datasets for predicting gender}
\usage{
install_genderdata_package()
}
\description{
If the genderdata package is not installed, install it from GitHub using
devtools. If it is not up to date, reinstall it.
}

