% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ipums_usa}
\alias{ipums_usa}
\title{IPUMS US census national names dataset}
\format{A data frame with c. 1.3m observations and 4 variables}
\source{
Steven Ruggles, J. Trent Alexander, Katie Genadek, Ronald Goeken, Matthew B. Schroeder, and Matthew Sobek. Integrated Public Use Microdata Series: Version 5.0 [Machine-readable database]. Minneapolis, MN: Minnesota Population Center [producer and distributor], 2010.
}
\description{
A data set containing the number of instances of male and female names born
in the years 1789 to 1940 for people included in the decennial IPUMS censuses
Numbers. IPUMS includes 1% and 5% samples for each year: for every birth year, counts
from all relevant censuses with people under 62 years old are included.
Contributed by Ben Schmidt.
}
\keyword{datasets}

