% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{kantrowitz}
\alias{kantrowitz}
\title{Katrowitz names corpus}
\format{A data frame with 7,579 observations and 2 variables}
\source{
Mark Kantrowitz and Bill Ross,
  \url{http://www.cs.cmu.edu/afs/cs/project/ai-repository/ai/areas/nlp/corpora/names/0.html}
}
\description{
A data set containing 7,579 unique names compiled into two lists of male and
female names by Mark Kantrowitz and Bill Ross in 1991, also used in Python's
Natural Language Toolkit.
}
\keyword{datasets}

