% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_genderdata_package}
\alias{check_genderdata_package}
\title{Check whether to install data for gender function and install if necessary}
\usage{
check_genderdata_package()
}
\description{
If the genderdata package is not installed, install it from GitHub using
devtools. If it is not up to date, reinstall it.
}

