% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gender-package.r
\docType{package}
\name{gender-package}
\alias{gender-package}
\title{Gender: predict gender by name from historical data}
\description{
Gender: predict gender from names using historical data
}
\details{
Encodes gender based on names and dates of birth, using U.S. Census or Social
Security data sets. Requires separate downlaod of datasets, which should be
done automatically and can be done manually by running
\code{install_genderdata_package()}.
}
\author{
\email{lincoln@lincolnmullen.com}
}

