% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{gender}
\alias{gender}
\alias{gender-package}
\title{Gender: find gender by name and date}
\usage{
gender(data, years = c(1932, 2012), method = "ssa", certainty = TRUE)
}
\arguments{
\item{data}{A character string of a first name or a data frame with a column
named \code{name} with a character vector containing first names. The names
must all be lowercase.}

\item{years}{This argument can be either a single year, a range of years in
the form \code{c(1880, 1900)}, or the value \code{TRUE}. If no value is
specified, then the names will be looked up for the period 1932 to 2012. If
a year or range of years is specified, then the names will be looked up for
that period. If the value is \code{TRUE}, then the function will look for
a column in the data frame named \code{year} containing an integer vector
of the year of birth associated with each name. This permits you to do a
precise lookup for each person in your data set. Dates may range from 1880
to 2012; if earlier or later dates are included in a column in the data
frame, they will not be matched.}

\item{method}{This value can be either \code{"ssa"}, in which case the
function will look up names based on Social Security Administration name
data, or \code{"kantrowitz"}, in which case the function will use the
Kantrowitz corpus of male and female names.}

\item{certainty}{A boolean value, which determines whether or not to return
the proportion of male and female uses of names in addition to determining
the gender of names.}
}
\description{
Gender: find gender by name and date

This function looks up the gender of either a single first name or of a
column of first names in a data frame. Optionally it can take a year, a
range of years, or a column of years in the data frame to take into account
variation in the use of names over time.
}
\details{
Encodes gender based on names and dates of birth, using the Social
Security Administration's data set of first names by year and state. By
using the SSA data instead of lists of male and female names, this package
is able to more accurately guess the gender of a name, and it is able to
report the probability that a name was male or female. Based on an technique
devised by Cameron Blevins.
}
\examples{
library(dplyr)
gender("madison")
gender("madison", years = c(1900, 1985))
gender("madison", years = 1985)
gender(sample_names_data)
gender(sample_names_data, years = TRUE)
gender(sample_names_data, certainty = FALSE)
gender(sample_names_data, method = "kantrowitz")
}
\author{
\email{lincoln@lincolnmullen.com}
}
\keyword{gender}

