% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gender-ipums-usa.R
\name{gender_ipums_usa}
\alias{gender_ipums_usa}
\title{Find the gender of first names using U.S. Census data}
\usage{
gender_ipums_usa(name, years, certainty)
}
\arguments{
\item{name}{A character string of a first name. Case insensitive.}

\item{years}{This argument can be either a single year, a range of years in
the form \code{c(1880, 1900)}. If no value is specified, then the names
will be looked up for the period 1789 to 1930. If a year or range of years
is specified, then the names will be looked up for that period. Acceptable
years range from 1789 to 1930.}

\item{certainty}{A boolean value, which determines whether or not to return
the proportion of male and female uses of names in addition to determining
the gender of names.}
}
\description{
This internal function implements the \code{method = "ipums"} option of
\code{\link{gender}}. See that function for documentation.
}

