% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gender_kantrowitz}
\alias{gender_kantrowitz}
\title{Find the gender of frst names using Kantrowitz names corpus}
\usage{
gender_kantrowitz(data)
}
\description{
This internal function implements the \code{method = "kantrowitz"} option of
\code{\link{gender}}. See that function for documentation.
}

