% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gender-kantrowitz.R
\name{gender_kantrowitz}
\alias{gender_kantrowitz}
\title{Find the gender of frst names using Kantrowitz names corpus}
\usage{
gender_kantrowitz(name)
}
\arguments{
\item{name}{A character string of a first name.}
}
\value{
A list or (for multiple names) a list of lists containing the name
  property and the predicted gender property.
}
\description{
This internal function implements the \code{method = "kantrowitz"} option of
\code{\link{gender}}. See that function for documentation.
}

