% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gender_ssa}
\alias{gender_ssa}
\title{Find the gender of frst names using Social Security data}
\usage{
gender_ssa(data, years, certainty)
}
\description{
This internal function implements the \code{method = "ssa"} option of
\code{\link{gender}}. See that function for documentation.
}

