% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gender-ssa.R
\name{get_correction_factors}
\alias{get_correction_factors}
\title{Calculate the correction factors for a year or range of years}
\usage{
get_correction_factors(years)
}
\arguments{
\item{years}{The range of years. This value will be passed to it by the
gender_ssa function.}
}
\description{
The SSA data is skewed by gender, especially for years before 1935. This
internal function figures out the factor by which the gender ratio should be
multiplied in order to assume that the ratio of male to female births for
those years was 1:1.
}

