% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install-genderdata-package.R
\name{install_genderdata_package}
\alias{install_genderdata_package}
\title{Install the genderdata package after checking with the user}
\usage{
install_genderdata_package()
}
\description{
Install the genderdata package after checking with the user
}

